+++ tests/mybatis-groovy/src/test/groovy/example/micronaut/mappers/GenreMapper.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package example.micronaut.genre
+package example.micronaut.mappers
 
 import example.micronaut.domain.Genre
 import org.apache.ibatis.annotations.Delete
@@ -22,6 +22,8 @@ import org.apache.ibatis.annotations.Options
 import org.apache.ibatis.annotations.Param
 import org.apache.ibatis.annotations.Select
 import org.apache.ibatis.annotations.Update
+
+import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Positive
 import jakarta.validation.constraints.PositiveOrZero
@@ -29,32 +31,31 @@ import jakarta.validation.constraints.PositiveOrZero
 interface GenreMapper {
 
     @Select("select * from genre where id=#{id}")
-    fun findById(id: Long): Genre?
+    Genre findById(long id)
 
     @Insert("insert into genre(name) values(#{name})")
     @Options(useGeneratedKeys = true, keyProperty = "id")
-    fun save(genre: Genre)
+    void save(Genre genre)
 
     @Delete("delete from genre where id=#{id}")
-    fun deleteById(id: Long)
+    void deleteById(long id)
 
-    @Update("update genre set name=#{name} where id=#{id}")
-    fun update(@Param("id") id: Long, @Param("name") name: String?)
+    @Update('update genre set name=#{name} where id=#{id}')
+    void update(@Param("id") long id, @Param("name") String name)
 
     @Select("select * from genre")
-    fun findAll(): List<Genre>
+    List<Genre> findAll()
 
-    @Select("select * from genre order by \${sort} \${order}")
-    fun findAllBySortAndOrder(@Param("sort") @Pattern(regexp = "id|name") sort: String,
-                              @Param("order") @Pattern(regexp = "asc|ASC|desc|DESC") order: String): List<Genre>
+    @Select('select * from genre order by ${sort} ${order}')
+    List<Genre> findAllBySortAndOrder(@Param("sort") @NotNull @Pattern(regexp = "id|name") String sort,
+                                      @Param("order") @NotNull @Pattern(regexp = "asc|ASC|desc|DESC") String order)
 
-    @Select("select * from genre order by \${sort} \${order} limit \${offset}, \${max}")
-    fun findAllByOffsetAndMaxAndSortAndOrder(@Param("offset") @PositiveOrZero offset: Int,
-                                             @Param("max") @Positive max: Int,
-                                             @Param("sort") @Pattern(regexp = "id|name") sort: String,
-                                             @Param("order") @Pattern(regexp = "asc|ASC|desc|DESC") order: String): List<Genre>
+    @Select('select * from genre order by ${sort} ${order} limit ${offset}, ${max}')
+    List<Genre> findAllByOffsetAndMaxAndSortAndOrder(@Param("offset") @PositiveOrZero int offset,
+                                                     @Param("max") @Positive int max,
+                                                     @Param("sort") @NotNull @Pattern(regexp = "id|name") String sort,
+                                                     @Param("order") @NotNull @Pattern(regexp = "asc|ASC|desc|DESC") String order)
 
-    @Select("select * from genre limit \${offset}, \${max}")
-    fun findAllByOffsetAndMax(@Param("offset") @PositiveOrZero offset: Int,
-                              @Param("max") @Positive max: Int): List<Genre>
+    @Select('select * from genre limit ${offset}, ${max}')
+    List<Genre> findAllByOffsetAndMax(@Param("offset") @PositiveOrZero int offset, @Param("max") @Positive int max)
 }
