+++ tests/mybatis-java/src/test/java/example/micronaut/genre/GenreRepositoryImpl.java
@@ -17,6 +17,7 @@ package example.micronaut.genre;
 
 import example.micronaut.ListingArguments;
 import example.micronaut.domain.Genre;
+import example.micronaut.mappers.GenreMapper;
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.validation.Validated;
 import jakarta.inject.Singleton;
